+++ NEWS.md
@@ -1,4 +1,5 @@
 ## 5.2.0 - Unrleased
+* [MODINVOICE-251](https://issues.folio.org/browse/MODINVOICE-251) - Remove zipping mechanism for data-import event payloads and use cache for job profile snapshot
 
 ## 5.1.0 - Released